refactoring dataset and project module; extracting code into columns and tabularData module
