+++ api/src/tabular-data/tabular-data.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from '@nestjs/common';
+
+@Controller('tabular-data')
+export class TabularDataController {}
