+++ api/src/tabular-data/tabular-data.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class TabularDataService {}
